restrict access to edit/delete

+++ app/controllers/articles_controller.rb
@@ -29,22 +29,34 @@ class ArticlesController < ApplicationController
   # edit action submits to update action
   def edit
     # @article = Article.find(params[:id])
+    unless @article.user == current_user
+      flash[:alert] = 'You can only edit or delete your own article'
+      redirect_to root_path
+    end
   end
 
   def update
-    if @article.update(article_params)
-      flash[:success] = "Article has been updated"
-      redirect_to @article
+    unless @article.user == current_user
+      flash[:danger] = 'You can only edit or delete your own article'
     else
-      flash[:danger] = "Article has not been updated"
-      render :edit
+      if @article.update(article_params)
+        flash[:success] = "Article has been updated"
+        redirect_to @article
+      else
+        flash[:danger] = "Article has not been updated"
+        render :edit
+      end
     end
   end
 
   def destroy
-    if @article.destroy
+    unless @article.user == current_user
+      flash[:danger] = 'You can only edit or delete your own article'
+    else
+      if @article.destroy
       flash[:success] = "Article was successfully deleted"
       redirect_to articles_path
+      end
     end
   end
